+++ extension/src/Settings/CliSettingsLoader.ts
@@ -6,6 +6,7 @@ import {
   Settings,
   IAppManifest,
   ILaunchFile,
+  IAppSourceCopSettings,
 } from "./Settings";
 import { WorkspaceFile } from "./WorkspaceFile";
 import { settingsMap } from "./SettingsMap";
@@ -60,17 +61,24 @@ export function getLaunchSettings(workspaceFolderPath: string): LaunchSettings {
   return launchSettings;
 }
 
+export function getAppSourceCopSettings(
+  workspaceFolderPath: string
+): IAppSourceCopSettings {
+  const filePath = join(workspaceFolderPath, "AppSourceCop.json");
+
+  const appSourceCopSettingsJson = loadJson(filePath) as IAppSourceCopSettings;
+  if (!appSourceCopSettingsJson.mandatoryAffixes) {
+    appSourceCopSettingsJson.mandatoryAffixes = [];
+  }
+
+  return appSourceCopSettingsJson;
+}
+
 export function getAppManifest(workspaceFolderPath: string): AppManifest {
   const filePath = join(workspaceFolderPath, "app.json");
   const appSettings = loadJson(filePath) as IAppManifest;
 
-  const appManifest = new AppManifest(
-    workspaceFolderPath,
-    appSettings.id,
-    appSettings.name,
-    appSettings.publisher,
-    appSettings.version
-  );
+  const appManifest = new AppManifest(workspaceFolderPath, appSettings);
 
   return appManifest;
 }